+++ packages/authentication/modules/user/withUser/types.ts
@@ -1,4 +1,4 @@
-import { JWTContent } from '@baseapp-frontend/utils'
+import type { JWTContent } from '@baseapp-frontend/utils'
 
 export interface ComponentWithUser<TUser extends JWTContent> {
   user: TUser | null