+++ packages/authentication/utils/login.ts
@@ -1,22 +1,13 @@
-import { TokenTypes } from '@baseapp-frontend/utils'
-
-import {
+import type {
   LoginChangeExpiredPasswordRedirectResponse,
-  LoginJWTResponse,
   LoginMfaResponse,
   LoginResponse,
-  LoginSimpleTokenResponse,
 } from '../types/auth'
 
 export const isLoginMfaResponse = (data: LoginResponse): data is LoginMfaResponse => {
   const mfaKey: keyof LoginMfaResponse = 'method'
   return mfaKey in data
 }
-
-export const isJWTResponse = (
-  token: TokenTypes,
-  response?: LoginJWTResponse | LoginSimpleTokenResponse,
-): response is LoginJWTResponse => token === TokenTypes.jwt
 
 export const isLoginChangeExpiredPasswordRedirectResponse = (
   data: LoginResponse,