+++ packages/config/.eslintrc.js
@@ -14,11 +14,13 @@ module.exports = {
     },
   },
   ignorePatterns: [
+    '__tests__/',
+    '__mocks__/',
+    'dist/',
     '.eslintrc.js',
     'cypress.config.ts',
     'jest.config.js',
     'jest.config.ts',
-    '**/__mocks__/**',
     'next.config.js',
     'postcss.config.js',
     'tailwind.config.js',
@@ -39,6 +41,7 @@ module.exports = {
   rules: {
     'global-require': 0,
     '@emotion/jsx-import': 'error',
+    'no-html-link-for-pages': 0,
     '@emotion/no-vanilla': 'error',
     '@emotion/import-from-emotion': 'error',
     '@emotion/styled-import': 'error',