+++ packages/core/src/permissions/index.tsx
@@ -1,4 +1,0 @@
-export * from './utils'
-export * from './withPermissions'
-export * from './types.d'
-export * from './constants'