+++ packages/eslint-plugin/.prettierrc.js
@@ -1,9 +1,1 @@
-module.exports = {
-  singleQuote: true,
-  trailingComma: 'all',
-  semi: false,
-  printWidth: 100,
-  importOrder: ['^react$', '^@baseapp-frontend/(.*)$', '<THIRD_PARTY_MODULES>', '^[./]'],
-  importOrderSeparation: true,
-  importOrderSortSpecifiers: true,
-}
+module.exports = require('@baseapp-frontend/config/.prettierrc.js')