+++ packages/utils/constants/jwt.ts
@@ -1,0 +1,2 @@
+export const ACCESS_KEY_NAME = 'Authorization'
+export const REFRESH_KEY_NAME = 'Refresh'